+++ resources/lang/en/common.php
@@ -3,7 +3,9 @@
 return [
     'change' => 'change',
     'confirm' => 'confirm',
+    'date_format' => 'date format',
     'ip_address' => 'ip address',
+    'language' => 'language',
     'last_update' => 'last update',
     'new' => 'new',
     'new_f' => 'new',
@@ -11,6 +13,9 @@ return [
     'old_f' => 'old',
     'old_password_doesnt_match' => 'old password is wrong',
     'operation_completed_successfully' => 'operation completed successfully',
+    'operation_failed' => 'operation failed',
     'settings' => 'settings',
+    'timezone' => 'timezone',
     'update' => 'update',
+    'user' => 'user',
 ];