+++ resources/lang/it/common.php
@@ -3,7 +3,9 @@
 return [
     'change' => 'cambio',
     'confirm' => 'conferma',
+    'date_format' => 'formato data',
     'ip_address' => 'indirizzo IP',
+    'language' => 'lingua',
     'last_update' => 'ultimo aggiornamento',
     'new' => 'nuovo',
     'new_f' => 'nuova',
@@ -11,6 +13,9 @@ return [
     'old_f' => 'vecchia',
     'old_password_doesnt_match' => 'la vecchia password è errata',
     'operation_completed_successfully' => 'operazione completata con successo',
+    'operation_failed' => 'operazione fallita',
     'settings' => 'impostazioni',
+    'timezone' => 'fuso orario',
     'update' => 'aggiorna',
+    'user' => 'utente',
 ];