+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+# 0.3.0
++ Add settings page for user
++ Retrieve user settings
++ Add date format user option
++ Add language user option
+
 # 0.2.0
-# 29-03-2019
 + Add bootstrap layout
 + Add Public IP on dashboard
 + Enable change password
 
 # 0.1.0
-## 24-03-2019
 + First boot
 + Add get public IP